refactor: move forms data to constants file
