+++ examples/app/payment-method-registration/page.tsx
@@ -1,80 +1,15 @@
 'use client';
 
-import type { RJSFSchema } from '@rjsf/utils';
-
 import { useState } from 'react';
 import { SchemaForm } from '@concepta/react-material-ui/dist';
-import { Text } from '@concepta/react-material-ui';
-import Box from '@mui/material/Box';
-import Container from '@mui/material/Container';
-import Card from '@mui/material/Card';
-import Button from '@mui/material/Button';
+import { Box, Container, Card, Button } from '@mui/material';
 
 import {
-  CustomTextFieldWidget,
-  CustomEmailFieldWidget,
-} from '@concepta/react-material-ui/dist/styles/CustomWidgets';
-import { AdvancedProperty } from '@concepta/react-material-ui/dist/components/SchemaForm/types';
-
-export interface PaymentMethodRegistrationFormData {
-  fullName: string;
-  email: string;
-  phone: string;
-  address: string;
-  city: number | null;
-  state: number | null;
-  zipCode: string;
-  cardNumber: string;
-  expirationDate: string;
-  cvc: string;
-}
-
-export const schema: RJSFSchema = {
-  type: 'object',
-  required: [
-    'fullName',
-    'email',
-    'phone',
-    'address',
-    'city',
-    'state',
-    'zipCode',
-    'expirationDate',
-    'cvc',
-  ],
-  properties: {
-    fullName: { type: 'string', title: 'Full name', minLength: 3 },
-    email: { type: 'string', title: 'Email', minLength: 3 },
-    phone: { type: 'string', title: 'Phone', minLength: 3 },
-    address: { type: 'string', title: 'Address', minLength: 3 },
-    city: { type: 'string', title: 'City', minLength: 3 },
-    state: { type: 'string', title: 'State', minLength: 3 },
-    zipCode: { type: 'string', title: 'Zip code', minLength: 3 },
-    cardNumber: { type: 'string', title: 'Card number', minLength: 3 },
-    expirationDate: { type: 'string', title: 'Expiration date', minLength: 3 },
-    cvc: { type: 'string', title: 'CVC', minLength: 3 },
-  },
-};
-
-export const advancedProperties: Record<string, AdvancedProperty> = {
-  email: {
-    type: 'string',
-  },
-  username: {
-    type: 'string',
-  },
-  password: {
-    type: 'password',
-  },
-  passwordConfirmation: {
-    type: 'password',
-  },
-};
-
-export const widgets = {
-  TextWidget: CustomTextFieldWidget,
-  EmailWidget: CustomEmailFieldWidget,
-};
+  type PaymentMethodRegistrationFormData,
+  schema,
+  advancedProperties,
+  widgets,
+} from './constants';
 
 const PaymentMethodRegistration = () => {
   const [formData, setFormData] = useState<PaymentMethodRegistrationFormData>({
@@ -88,6 +23,7 @@ const PaymentMethodRegistration = () => {
     cardNumber: '',
     expirationDate: '',
     cvc: '',
+    saveAsDefault: true,
   });
 
   const handleSubmit = async (values: PaymentMethodRegistrationFormData) => {
@@ -97,18 +33,9 @@ const PaymentMethodRegistration = () => {
 
   return (
     <Container maxWidth="xs" sx={{ textAlign: 'center', padding: '48px 0' }}>
-      <Text
-        variant="h4"
-        fontFamily="Inter"
-        fontSize={30}
-        fontWeight={800}
-        mt={1}
-        gutterBottom
-      >
-        Payment method registration
-      </Text>
+      <h1>Payment method registration</h1>
 
-      <Card sx={{ marginTop: '26px', padding: '24px' }}>
+      <Card sx={{ marginTop: '48px', padding: '24px' }}>
         <SchemaForm.Form
           schema={schema}
           formData={formData}
